Add the first unit tests
