+++ phpunit.xml.dist
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<phpunit backupGlobals="false"
+         convertErrorsToExceptions="true"
+         convertWarningsToExceptions="true"
+         convertNoticesToExceptions="true"
+         mapTestClassNameToCoveredClassName="true"
+         bootstrap="vendor/autoload.php"
+         strict="true"
+         verbose="true"
+         colors="true">
+    <testsuites>
+        <testsuite name="Sitemap-Generator-Crawler">
+            <directory>./tests/</directory>
+        </testsuite>
+    </testsuites>
+</phpunit>
